accessible/tests/mochitest/attributes/test_obj_group_tree.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL tree attributes tests">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../treeview.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../attributes.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Test

    function doTestAttrs()
    {
      var treeNode = getNode("tree");

      var tree = getAccessible(treeNode);
      var treeitem1 = tree.firstChild.nextSibling;
      testGroupAttrs(treeitem1, 1, 4, 1);

      var treeitem2 = treeitem1.nextSibling;
      testGroupAttrs(treeitem2, 2, 4, 1);

      var treeitem3 = treeitem2.nextSibling;
      testGroupAttrs(treeitem3, 1, 2, 2);

      var treeitem4 = treeitem3.nextSibling;
      testGroupAttrs(treeitem4, 2, 2, 2);

      var treeitem5 = treeitem4.nextSibling;
      testGroupAttrs(treeitem5, 3, 4, 1);

      var treeitem6 = treeitem5.nextSibling;
      testGroupAttrs(treeitem6, 4, 4, 1);

      SimpleTest.finish();
    }

    function doTest()
    {
      var treeNode = getNode("tree");
      waitForEvent(EVENT_REORDER, treeNode, doTestAttrs);
      treeNode.view = new nsTreeTreeView();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=503727"
         title="Reorganize implementation of XUL tree accessibility">
        Mozilla Bug 503727
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <tree id="tree" flex="1">
        <treecols>
          <treecol id="col" flex="1" primary="true" label="column"/>
        </treecols>
        <treechildren/>
      </tree>

      <vbox id="debug"/>
    </vbox>
  </hbox>

</window>
